requirements.txt
Flask==0.12.2
Jinja2==2.10
MarkupSafe==1.0
Savoir==1.0.6
Werkzeug==0.14.1
click==6.7
get==0.0.39
itsdangerous==0.24
post==0.0.26
public==0.0.65
query-string==0.0.28
request==0.0.26
virtualenv==15.1.0
